% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_project.R
\name{text_project}
\alias{is.global}
\alias{text_project}
\title{Text Project Template}
\usage{
text_project(path = "new", open = is.global(2))

is.global(n = 1)
}
\arguments{
\item{path}{The path (including project name) for the project.}

\item{open}{logical. If \code{TRUE} the project will be opened in RStudio.
The default is to test if new_project is being used in the global environment,
if it is then the project directory will be opened.}
}
\description{
Create a template directory of subdirectories and files for a text project.
}
\examples{
\dontrun{
text_project()
}
}
\keyword{project}

